Add files via upload
The creatinine phosphokinase (CPK) states the level of the CPK enzyme in blood measured in micrograms per liter (mcg/L). When a muscle tissue gets damaged, CPK flows into the blood. Therefore, high levels of CPK in the blood of a patient might indicate heart failure or injury. This histogram is skewed right as most of the data fall between 0 - 1000 mcg/L.